# contracts/product/v1/openapi.yaml
openapi: 3.1.2
...
info:
  title: Cyrene Exchange Product API
  version: 1.0.0
  license:
    name: Apache-2.0
paths:
  /api/v1/gateway-endpoints:
    post:
      operationId: createGatewayEndpoint
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEndpointRequest'
      responses:
        '201':
          description: GatewayEndpoint created or replayed
          content:
            application/json:
              schema:
                $ref: ./gateway-endpoint.schema.json
        '409':
          $ref: '#/components/responses/Problem'
    get:
      operationId: listGatewayEndpoints
      responses:
        '200':
          description: Persisted GatewayEndpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: ./gateway-endpoint.schema.json
  /api/v1/gateway-endpoints/{endpointId}:
    get:
      operationId: getGatewayEndpoint
      parameters:
      - $ref: '#/components/parameters/EndpointId'
      responses:
        '200':
          description: GatewayEndpoint
          content:
            application/json:
              schema:
                $ref: ./gateway-endpoint.schema.json
        '404':
          $ref: '#/components/responses/Problem'
  /api/v1/gateway-endpoints/{endpointId}/actions/disable:
    post:
      operationId: disableGatewayEndpoint
      parameters:
      - $ref: '#/components/parameters/EndpointId'
      responses:
        '200':
          description: Disabled GatewayEndpoint
          content:
            application/json:
              schema:
                $ref: ./gateway-endpoint.schema.json
        '404':
          $ref: '#/components/responses/Problem'
  /api/v1/gateway-routes:
    post:
      operationId: createGatewayRoute
      description: Legacy organization-unscoped write. Organization-bound credentials must use scoped Workspace routes.
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRouteRequest'
      responses:
        '201':
          description: GatewayRoute created or replayed
          content:
            application/json:
              schema:
                $ref: ./gateway-route.schema.json
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
    get:
      operationId: listGatewayRoutes
      description: Lists only rows whose organization attribution is unknown.
      responses:
        '200':
          description: Legacy GatewayRoutes without organization attribution
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: ./gateway-route.schema.json
  /api/v1/gateway-routes/{routeId}:
    get:
      operationId: getGatewayRoute
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Legacy GatewayRoute without organization attribution
          content:
            application/json:
              schema:
                $ref: ./gateway-route.schema.json
        '404':
          $ref: '#/components/responses/Problem'
  /api/v1/workspace/gateway-routes:
    get:
      operationId: listWorkspaceGatewayRoutes
      security:
      - WorkspaceControlBearer: []
      responses:
        '200':
          description: GatewayRoutes assigned to the authenticated organization and workspace
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: ./gateway-route.schema.json
        '403':
          $ref: '#/components/responses/Problem'
  /api/v1/workspace/gateway-route-drafts:
    post:
      operationId: createWorkspaceGatewayRouteDraft
      security:
      - WorkspaceControlBearer: []
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRouteDraftRequest'
      responses:
        '201':
          description: DRAFT created for the authenticated organization and workspace
          content:
            application/json:
              schema:
                $ref: ./gateway-route.schema.json
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          description: Request validation failed
  /api/v1/api-keys:
    post:
      operationId: createApiKey
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: API key metadata; the secret is present only when created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedApiKey'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
    get:
      operationId: listApiKeys
      responses:
        '200':
          description: API key metadata without secret material
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
        '403':
          $ref: '#/components/responses/Problem'
  /api/v1/api-keys/{apiKeyId}:
    get:
      operationId: getApiKey
      parameters:
      - name: apiKeyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: API key metadata without secret material
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '404':
          $ref: '#/components/responses/Problem'
  /api/v1/api-keys/{apiKeyId}/actions/revoke:
    post:
      operationId: revokeApiKey
      parameters:
      - name: apiKeyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Revoked API key metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '404':
          $ref: '#/components/responses/Problem'
...
  /api/v1/gateway-route-drafts:
    post:
      summary: Create Draft
      operationId: create_draft_api_v1_gateway_route_drafts_post
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRouteDraftRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayRoute'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateway-route-drafts/{routeId}:
    patch:
      summary: Edit Draft
      operationId: edit_draft_api_v1_gateway_route_drafts__routeId__patch
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Routeid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditRouteDraftRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayRoute'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gateway-route-drafts/{routeId}/actions/confirm:
    post:
      summary: Confirm Draft
      operationId: confirm_draft_api_v1_gateway_route_drafts__routeId__actions_confirm_post
      parameters:
      - name: routeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Routeid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmRouteRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GatewayRoute'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  securitySchemes:
    WorkspaceControlBearer:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 200
    EndpointId:
      name: endpointId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Problem:
      description: RFC 9457 control-plane error
      content:
        application/problem+json:
          schema:
            $ref: ./generated/common/problem-details.schema.json
  schemas:
    CreateEndpointRequest:
      type: object
      required:
      - name
      - publicBaseUrl
      - authPolicyRef
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        publicBaseUrl:
          type: string
          format: uri
        authPolicyRef:
          type: string
          minLength: 1
          maxLength: 500
      additionalProperties: false
    CreateRouteRequest:
      type: object
      required:
      - endpointId
      - modelPattern
      - targetBindingId
      - priority
      properties:
        endpointId:
          type: string
          format: uuid
        modelPattern:
          type: string
          minLength: 1
          maxLength: 200
        targetBindingId:
          type: string
          minLength: 1
          maxLength: 300
        targetModel:
          type: string
          minLength: 1
          maxLength: 200
        priority:
          type: integer
          minimum: 0
          maximum: 10000
      additionalProperties: false
...
    ConfirmRouteRequest:
      properties:
        resourceVersion:
          type: integer
          minimum: 1.0
          title: Resourceversion
      additionalProperties: false
      type: object
      required:
      - resourceVersion
      title: ConfirmRouteRequest
      description: Explicit confirmation of the exact inspected draft version. | 确认草稿版本。
    CreateRouteDraftRequest:
      properties:
        endpointId:
          type: string
          format: uuid
          title: Endpointid
        modelPattern:
          type: string
          maxLength: 200
          minLength: 1
          title: Modelpattern
        targetBindingId:
          type: string
          maxLength: 300
          minLength: 1
          title: Targetbindingid
        targetModel:
          type: string
          maxLength: 200
          minLength: 1
          title: Targetmodel
        priority:
          type: integer
          maximum: 10000.0
          minimum: 0.0
          title: Priority
        source:
          $ref: '#/components/schemas/RouteSource'
      additionalProperties: false
      type: object
      required:
      - endpointId
      - modelPattern
      - targetBindingId
      - targetModel
      - priority
      - source
      title: CreateRouteDraftRequest
      description: Explicit Send to creates a draft, never an active route. | 发送到只创建草稿。
    EditRouteDraftRequest:
      properties:
        resourceVersion:
          type: integer
          minimum: 1.0
          title: Resourceversion
        modelPattern:
          type: string
          maxLength: 200
          minLength: 1
          title: Modelpattern
        targetBindingId:
          type: string
          maxLength: 300
          minLength: 1
          title: Targetbindingid
        targetModel:
          type: string
          maxLength: 200
          minLength: 1
          title: Targetmodel
        priority:
          type: integer
          maximum: 10000.0
          minimum: 0.0
          title: Priority
      additionalProperties: false
      type: object
      required:
      - resourceVersion
      - modelPattern
      - targetBindingId
      - targetModel
      - priority
      title: EditRouteDraftRequest
      description: Editable route intent with optimistic concurrency. | 草稿修改与版本检查。
...
    CreateApiKeyRequest:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        expiresAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        modelScope:
          items:
            type: string
          maxItems: 100
          type: array
          title: Modelscope
      additionalProperties: false
      type: object
      required:
      - name
      title: CreateApiKeyRequest
      description: Server-generated gateway key request. | 服务端生成网关密钥请求。
    ApiKey:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
        credentialRef:
          type: string
          pattern: ^api-key://[0-9a-f-]{36}$
          title: Credentialref
        actorId:
          type: string
          maxLength: 300
          minLength: 1
          title: Actorid
        workspaceId:
          type: string
          maxLength: 300
          minLength: 1
          title: Workspaceid
        state:
          $ref: '#/components/schemas/ApiKeyState'
        modelScope:
          items:
            type: string
          maxItems: 100
          type: array
          title: Modelscope
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
        expiresAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        revokedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        resourceVersion:
          type: integer
          minimum: 1.0
          title: Resourceversion
      additionalProperties: false
      type: object
      required:
      - id
      - name
      - credentialRef
      - actorId
      - workspaceId
      - state
      - createdAt
      - updatedAt
      - resourceVersion
      title: ApiKey
      description: Gateway credential metadata; the secret is never persisted. | 密钥元数据。
    CreatedApiKey:
      allOf:
      - $ref: '#/components/schemas/ApiKey'
      - properties:
          secret:
            anyOf:
            - type: string
            - type: 'null'
            title: Secret
        type: object
        title: CreatedApiKey
      description: Creation response carrying the one-time secret. | 仅创建响应携带一次性密钥。
...
    GatewayRoute:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        endpointId:
          type: string
          format: uuid
          title: Endpointid
        state:
          $ref: '#/components/schemas/RouteState'
          default: ACTIVE
        modelPattern:
          type: string
          maxLength: 200
          minLength: 1
          title: Modelpattern
        targetBindingId:
          type: string
          maxLength: 300
          minLength: 1
          title: Targetbindingid
        targetCapabilityType:
          type: string
          const: model.provider.v1
          title: Targetcapabilitytype
          default: model.provider.v1
        targetModel:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Targetmodel
        priority:
          type: integer
          maximum: 10000.0
          minimum: 0.0
          title: Priority
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
        resourceVersion:
          type: integer
          minimum: 1.0
          title: Resourceversion
        source:
          anyOf:
          - $ref: '#/components/schemas/RouteSource'
          - type: 'null'
        createdBy:
          anyOf:
          - type: string
          - type: 'null'
          title: Createdby
        workspaceId:
          anyOf:
          - type: string
          - type: 'null'
          title: Workspaceid
      additionalProperties: false
      type: object
      required:
      - id
      - endpointId
      - modelPattern
      - targetBindingId
      - priority
      - createdAt
      - updatedAt
      - resourceVersion
      title: GatewayRoute
      description: Product route to an opaque provider capability binding. | 指向不透明能力绑定的路由。
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    RouteSource:
      properties:
        product:
          type: string
          const: reactor
          title: Product
          default: reactor
        resourceUri:
          type: string
          maxLength: 2000
          pattern: ^https?://
          title: Resourceuri
        resourceVersion:
          type: integer
          minimum: 1.0
          title: Resourceversion
        artifactDigest:
          type: string
          pattern: ^sha256:[0-9a-f]{64}$
          title: Artifactdigest
        modelVersionId:
          anyOf:
          - type: string
            pattern: ^model-version://sha256/[0-9a-f]{64}$
          - type: 'null'
          default: null
          title: Modelversionid
      additionalProperties: false
      type: object
      required:
      - resourceUri
      - resourceVersion
      - artifactDigest
      title: RouteSource
      description: Versioned reference to an existing Reactor Endpoint. | 来源端点引用。
...